Move Options::default() to impl Default
This implements the `Default` trait for `Options`, which was as simple
as moving our existing `default()` method to inside of an impl
block. This should make it easier for people to override just one or two
of the fields on `Options`, without having to specify all fields through
the `new()` method.