Make DogstatsdError-type publicly available
When consuming this crate's API it is sometimes desirable to be able
to access the internal error type in other code, for example when
setting up error chains or writing type-specific handler logic.

This change exposes the error type by publicly using it from the
`errors` module and adds (admittedly very simple) documentation on the
error type itself.